third_party/WebKit/LayoutTests/fast/css/recalc-optgroup-inherit.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<select id="s">
    <optgroup id="g" style="color:red">
        <option id="o">Green</option>
    </optgroup>
</select>
<script>
description("Check that OPTION inherits the modified OPTGROUP color.");

document.body.offsetTop; // force style recalc.
document.getElementById("g").style.color = "green";
shouldBe("getComputedStyle(document.getElementById('o'), null).color", "'rgb(0, 128, 0)'");
</script>
