third_party/WebKit/LayoutTests/fast/css/recalc-optgroup-inherit-expected.txt

Check that OPTION inherits the modified OPTGROUP color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.getElementById('o'), null).color is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
